Rename to SingletonParameterResolver
Also fix resolution of multiple @New resources in a single method.
